requirements.txt
openai
streamlit
python-dotenv
langchain
langchain-openai
black
isort
llama-index
#llama-index-llms-openai
#llama-index-program-openai
#llama-index-agent-openai
llama-index-core 
llama-index-llms-openai
psycopg2-binary
boto3
fastapi
uvicorn
pydantic
mysql-connector-python
langchain-chroma